# docs/swagger.yaml
swagger: "2.0"
host: localhost:8080
basePath: /billing
info:
  description: "Healtcare system Billing APIs."
  version: "1.0.0"
  title: "Health care system - billing"
  contact:
    email: "cs62@njit.edu"
tags:
- name: "codes"
  description: "Everything about procedures and rates"
- name: "claims"
  description: "Billing claims"
schemes:
- "https"
- "http"
paths:
  /icd10codes:
    get:
      tags:
      - "codes"
      summary: "List the range of ICD-10 codes to search from"
      description: "Returns the range search list of ICD-10 codes"
      operationId: "getICD10Codes"
      responses:
        "405":
          description: "Invalid input"
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ICD10"
  /icd10codes/{icd10codeId}:
    get:
      tags:
      - "codes"
      summary: "Returns the list of ICD-10 codes filtered by the {code}."
      description: "List the range of ICD-10 codes based on the code provided in the path. The code could be a range string provided as a code."
      operationId: "getICD10CodesBySearch"
      parameters:
      - name: "icd10codeId"
        in: "path"
        description: "Search string. This could be code range/group or the code. This has to be same as a code provided in above API or this API"
        required: true
        type: "string"
      responses:
        "405":
          description: "Invalid input"
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ICD10"
  /cptCodes:
    get:
      tags:
      - "codes"
      summary: "List all the CPT codes with the grouping"
      description: "Returns list of CPT codes with grouping"
      operationId: "getCPTCodes"
      responses:
        "405":
          description: "Invalid input"
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CPTGroup"
  /cptCodes/{groupId}:
    get:
      tags:
      - "codes"
      summary: "List all the CPT codes for the given group id."
      description: "Returns all the CPT codes for the given group id."
      operationId: "getCPTCodesBySearch"
      parameters:
      - name: "groupId"
        in: "path"
        description: "Group id for the code"
        required: true
        type: "string"
      responses:
        "405":
          description: "Invalid input"
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CPT"
  /cptCodes/rates:
    get:
      tags:
      - "codes"
      summary: "Rate of a procedure."
      description: "Returns the rate of a procedure with id {cptCodesId}."
      operationId: "procedureRates"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CPTCodeRate"
        "400":
          description: "Invalid status value"
  /cptCodes/rates/{cptCodesId}:
    get:
      tags:
      - "codes"
      summary: "Rate of a procedure."
      description: "Returns the rate of a procedure with id {cptCodesId}."
      operationId: "procedureRate"
      parameters:
      - name: "cptCodesId"
        in: "path"
        description: "Procedure Id"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CPTCodeRate"
        "400":
          description: "Invalid status value"
  /claims:
    get:
      tags:
      - "claims"
      summary: "Returns list of all claims"
      description: "Returns list of all claims"
      operationId: "getClaims"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Claim"
        "400":
          description: "Invalid tag value"
    post:
      tags:
      - "claims"
      summary: "Creates claim"
      description: "Creates claim based on the ICD and CPT codes provided for a given patient"
      operationId: "createClaim"
      parameters:
        - in: body
          name: "CreateClaimRequest"
          schema:
            $ref: "#/definitions/CreateClaimRequest"
      responses:
        "405":
          description: "Invalid input"
        "200":
          description: "Successful"
  /claims/{claimId}:
    get:
      tags:
      - "claims"
      summary: "Find claim by ID"
      description: "Returns a single claim"
      operationId: "getClaimById"
      parameters:
      - name: "claimId"
        in: "path"
        description: "ID of claim that needs to be updated"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Claim"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Pet not found"
    put:
      tags:
      - "claims"
      summary: "Updates a claim"
      description: ""
      operationId: "updateClaim"
      parameters:
      - name: "claimId"
        in: "path"
        description: "ID of pet that needs to be updated"
        required: true
        type: "string"
      responses:
        "405":
          description: "Invalid input"
    delete:
      tags:
      - "claims"
      summary: "Deletes a claim"
      description: ""
      operationId: "deleteClaim"
      parameters:
      - name: "claimId"
        in: "path"
        description: "ID of pet that needs to be updated"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Not found"
  /claims/{claimId}/actions/process:
    post:
      tags:
        - "claims"
      summary: "Process a claim"
      description: "Process a claim"
      operationId: "processClaim"
      parameters:
      - name: "claimId"
        in: "path"
        description: "ID of pet that needs to be updated"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Pet not found"
definitions:
  ICD10:
    type: "object"
    properties:
      id:
        type: "string"
        description: "object id"
      code:
        type: "string"
        description: "This could represent the ICD-10 code or a search string/grouping"
      description:
        type: "string"
        description: "Information about the code."
      subCodes:
        type: "array"
        items:
          "$ref": "#/definitions/ICD10"
  CPT:
    type: object
    properties:
      id:
        type: "string"
        description: "object id"
      code:
        type: "string"
        description: "CPT code"
      description:
        type: "string"
        description: "information about the code"
  CPTGroup:
    type: object
    properties:
      id:
        type: "string"
        description: "object id"
      type:
        type: "string"
        description: "Type of grouping"
      description:
        type: "string"
        description: "information about the grouping"
      additionalInfo:
        type: "string"
        description: "Any additional info about the grouping"
  Currency:
    type: object
    properties:
      identifier:
        type: string
      symbol:
        type: string
  CPTCodeRate:
    type: object
    properties:
      id:
        type: "string"
        description: "object id"
      cptCode:
        "$ref": "#/definitions/CPT"
      rate:
        type: integer
      currency:
        "$ref": "#/definitions/Currency"
  CreateClaimRequest:
    type: object
    properties:
      cptCodes:
        type: array
        items:
          "$ref": "#/definitions/CPTCodeRate"
      icd10Codes:
        type: array
        items:
          "$ref": "#/definitions/ICD10"
      patientId:
        type: string
  Claim:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "string"
        description: "claim id"
      cptCodes:
        type: "array"
        items:
          $ref: "#/definitions/CPT"
      icd10Codes:
        type: "array"
        items:
          $ref: "#/definitions/ICD10"
      patientId:
        type: "string"
      additionalInfo:
        type: "string"
      amount:
        type: "number"
      status:
        type: "string"
        description: "status of the claim"
        enum:
        - "initiated"
        - "accepted"
        - "rejected"
        - "denied"
        - "sent_for_adjudication"
        - "deleted"
      currency:
        "$ref": "#/definitions/Currency"
      timeCreated:
        type: "string"
        description: "ISO format date created"
      timeUpdated:
        type: "string"
        description: "ISO format date created"
        